rename extra_signals -> util_components
