+++ profile_bluesky/startup/instrument/devices/phaseplates.py
@@ -11,7 +11,7 @@ from ophyd import EpicsSignal, EpicsSignalRO, Signal, DerivedSignal
 from ophyd.pseudopos import pseudo_position_argument, real_position_argument
 from scipy.constants import speed_of_light, Planck
 from numpy import arcsin, pi, sin
-from .extra_signals import TrackingSignal
+from .util_components import TrackingSignal
 from ..session_logs import logger
 
 # This is here because PRDevice.select_pr has a micron symbol that utf-8
